Properly handle CUDA graphs.
In the previous version, the logic only triggered when CUDA graphs were
globally disabled. This is not necessary; we can instead selectively
disable graph support when a dynamic memcpy fusion is present.

This requires adding a new pass that annotates these fusions, so we can
detect them during command buffer scheduling.