Fix layout issues inside fusions.
Previously, there was a workaround in the fusion handler to set missing
layouts to the default layout. This is needed because layout assignment
will usually clear internal layouts in fusions. Since we no longer have
a copy of the module, we cannot override the layouts - instead, we
support missing layouts where necessary.

Also add a test for a layout-less 2d reduction.