Design sketch for loop TMA.
This currently contains a completely untested end-to-end pipeline.

There are lots of things left to do, but no major risks that I know
of.

An example of the top half of the pipeline can be found in
xla/backends/gpu/codegen/emitters/tests/loop/basic.hlo.

An example of the lower half of the pipeline can be found in
xla/codegen/emitters/transforms/tests/pipes.mlir.

I have not yet tried lowering to LLVM; that's almost certainly broken.